Extract workflow test utilities
Move duplicated act invocation and artifact-reading helpers into a
shared workflow test utility module. Keep explicit exit-code checks
documented and align rejected-source assertions with length checks.
